prototype/requirements.txt
# Simple prototype requirements
streamlit==1.28.0
tensorflow==2.13.0
tensorflow-lite==2.13.0
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2



